+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - If FFmpeg and/or FFprobe are missing from your system then you will be asked if you want Squash to download them for you
 - Added a recommendation for Windows Terminal if it isn't being used
+- Added `-v`/`--version` flag
 
 ## 1.0.1